refactor: define existing plugin as v2 plugin
