+++ src/v2-parts/printers.ts
@@ -1,7 +1,7 @@
 import type { AstPath, ParserOptions, Doc, Printer, Plugin } from 'prettier';
 import { format } from 'prettier';
 
-import { parseLineByLineAndReplace } from './core';
+import { parseLineByLineAndReplace } from '../core';
 
 function createPrinter(parserName: 'babel' | 'typescript'): Printer {
   function main(
