+++ src/v2-plugin.ts
@@ -1,8 +1,8 @@
 import type { Plugin } from 'prettier';
 
-import { options } from './options';
-import { parsers } from './parsers';
-import { printers } from './printers';
+import { options } from './v2-parts/options';
+import { parsers } from './v2-parts/parsers';
+import { printers } from './v2-parts/printers';
 
 const classnamesPlugin: Plugin = {
   parsers,
